Extract error formatter to module and allow custom formatter
