+++ lib/GraphQLAdapter.js
@@ -8,9 +8,11 @@ const { formatError } = require('./resolvers/error')
 function GraphQLAdapter(options) {
   const router = express.Router()
   const { services } = options
-  const defaults = { graphiql: true }
+  const defaults = { graphiql: true, errorFormatterImpl: '@cap-js/graphql/lib/errorFormatter' }
   const schema = generateSchema4(services)
   options = { ...defaults, ...options }
+
+  const errorFormatterImpl = require(options.errorFormatterImpl)
 
   router
     .use(express.json()) //> required by logger below
@@ -18,7 +20,9 @@ function GraphQLAdapter(options) {
 
   if (options.graphiql) router.use(graphiql)
 
-  router.use((req, res) => createHandler({ schema, context: { req, res }, formatError, ...options })(req, res))
+  router.use((req, res) =>
+    createHandler({ schema, context: { req, res, errorFormatterImpl }, formatError, ...options })(req, res)
+  )
 
   return router
 }
