+++ lib/resolvers/error.js
@@ -5,42 +5,6 @@ const { normalizeError } = require('@sap/cds/libx/_runtime/common/error/frontend
 const { GraphQLError } = require('graphql')
 const { IS_PRODUCTION } = require('../utils')
 
-const ALLOWED_PROPERTIES_IN_PRODUCTION = ['message', 'target', 'details']
-
-const _sanitizeProperty = (key, value, out) => {
-  if (key === 'i18nKey') {
-    out['code'] = value
-    return
-  }
-
-  if (IS_PRODUCTION) {
-    if (ALLOWED_PROPERTIES_IN_PRODUCTION.includes(key)) out[key] = value
-    return
-  }
-
-  if (key === 'stack') {
-    out['stacktrace'] = value.split('\n')
-    return
-  }
-
-  out[key] = value
-}
-
-const _errorFormatterFn = err => {
-  const error = {}
-
-  const properties = Object.keys(err).concat('message')
-
-  if (err.i18nKey) properties.push('i18nKey')
-
-  // No stack for outer error of multiple errors, since the stack is not meaningful
-  if (!err.details) properties.push('stack')
-
-  properties.forEach(k => _sanitizeProperty(k, err[k], error))
-
-  return error
-}
-
 const _reorderProperties = error => {
   let { code, message, details, stack } = error
   if (details) details = details.map(_reorderProperties)
@@ -48,7 +12,8 @@ const _reorderProperties = error => {
 }
 
 const _cdsToGraphQLError = (context, err) => {
-  let { error } = normalizeError(err, context.req, _errorFormatterFn)
+  const { req, errorFormatterImpl } = context
+  let { error } = normalizeError(err, req, errorFormatterImpl)
 
   // Ensure error properties are ordered nicely for the client
   error = _reorderProperties(error)
